Colocando cada componente com a sua responsabilidade de estado
