+++ semana7/integracao-api/src/Components/Inputs.js
@@ -1,7 +1,8 @@
 import React from 'react'
 import styled from 'styled-components'
+import axios from 'axios'
 
-
+//* **STYLED COMPONENTS** */
 const Container = styled.div`
   width:50%;
   height: 98vh;
@@ -21,39 +22,79 @@ const Wrapper = styled(Container)`
   font-size:1.5em;
   margin: 5% auto;
 `
+//--------------------------------------------------------------------------
 
-function Inputs(props) {
+//* **VARIAVEIS GLOBAIS JS** */
+const baseUrl = "https://us-central1-future4-users.cloudfunctions.net/api"
+const token = "danilo-sagan"
+//--------------------------------------------------------------------------
 
-  const onChangeInputValue = (event) => {
-    let newValue = event.target.value
-    let input = event.target.name
-    props.refreshValue(input, newValue)
+//* **COMPONENTE DE CLASSE** */
+class Inputs extends React.Component {
+  constructor(props) {
+    super(props)
+    this.state = {
+      nome: '',
+      email: ''
+    }
   }
-  return (
-    <Container>
-      <Wrapper>
-        <label for="nome">Nome:</label>
-        <input 
-          type="text" 
-          id="nome" 
-          name="nome"
-          value={props.nome}
-          onChange={onChangeInputValue} 
-        />
-        <label for="email">E-mail:</label>
-        <input 
-          type="email" 
-          id="email" 
-          name="email"
-          value={props.email}
-          onChange={onChangeInputValue}
-        />
-        <button onClick={props.sendData}>Enviar</button>
-      </Wrapper>
-      <button onClick={props.changePage}>Ir para página de lista</button>
-    </Container>
-  )
+
+  onChangeInputValue = (event) => {
+    this.setState({ [event.target.name]: event.target.value })
+  }
+
+  onDataSend = () => {
+    const dataToSend = {
+      name: this.state.nome,
+      email: this.state.email
+    }
+    
+    const request = axios.post(`${baseUrl}/users/createUser`, dataToSend, {
+      headers: {
+        "Content-Type": "application/json",
+        "api-token": token
+      }
+    })
+
+    request.then((response) => {
+      window.alert('Dados enviados com sucesso!')
+      this.setState({
+        nome: '',
+        email: ''
+      })
+    }).catch((error) => {
+      window.alert('Erro inesperado')
+    })
+  }
+
+  render() {
+    return (
+      <Container>
+        <Wrapper>
+          <label for="nome">Nome:</label>
+          <input
+            type="text"
+            id="nome"
+            name="nome"
+            value={this.state.nome}
+            onChange={this.onChangeInputValue}
+          />
+          <label for="email">E-mail:</label>
+          <input
+            type="email"
+            id="email"
+            name="email"
+            value={this.state.email}
+            onChange={this.onChangeInputValue}
+          />
+          <button onClick={this.onDataSend}>Enviar</button>
+        </Wrapper>
+        <button onClick={this.props.changePage}>Ir para página de lista</button>
+      </Container>
+    )
+  }
 }
+//----------------------------------------------------------------------------------
 
 
 export default Inputs
